Harmony branch

+++ server/seeds/cleanDB.js
@@ -1,13 +1,14 @@
-const models = require('../models');
-const db = require('../config/connection');
+const models = require("../models");
+const db = require("../config/connection");
+
 
 module.exports = async (modelName, collectionName) => {
   try {
-    let modelExists = await models[modelName].db.db.listCollections({
+    let modelExists = await models[modelName]?.db?.db?.listCollections({
       name: collectionName
     }).toArray()
 
-    if (modelExists.length) {
+    if (modelExists && modelExists.length) {
       await db.dropCollection(collectionName);
     }
   } catch (err) {
